+++ src/Config/Option/UrlShortener/ShortUrlModeConfigOption.php
@@ -5,16 +5,17 @@ declare(strict_types=1);
 namespace Shlinkio\Shlink\Installer\Config\Option\UrlShortener;
 
 use Shlinkio\Shlink\Installer\Config\Option\BaseConfigOption;
+use Shlinkio\Shlink\Installer\Config\Option\ConfigOptionMigratorInterface;
 use Symfony\Component\Console\Style\StyleInterface;
 
 use const PHP_EOL;
 
-class ShortUrlModeConfigOption extends BaseConfigOption
+class ShortUrlModeConfigOption extends BaseConfigOption implements ConfigOptionMigratorInterface
 {
     private const MODES = [
         'strict' => 'Short codes and custom slugs will be matched in a case-sensitive way ("foo" !== "FOO"). '
             . 'Generated short codes will include lowercase letters, uppercase letters and numbers.',
-        'loosely' => 'Short codes and custom slugs will be matched in a case-insensitive way ("foo" === "FOO"). '
+        'loose' => 'Short codes and custom slugs will be matched in a case-insensitive way ("foo" === "FOO"). '
             . 'Generated short codes will include only lowercase letters and numbers.',
     ];
 
@@ -35,5 +36,10 @@ class ShortUrlModeConfigOption extends BaseConfigOption
             $options,
             'strict',
         );
+    }
+
+    public function tryToMigrateValue(mixed $currentValue): mixed
+    {
+        return $currentValue === 'loosely' ? 'loose' : $currentValue;
     }
 }