+++ test/Config/Option/UrlShortener/ShortUrlModeConfigOptionTest.php
@@ -50,6 +50,22 @@ class ShortUrlModeConfigOptionTest extends TestCase
     public function provideChoices(): iterable
     {
         yield 'strict' => ['strict'];
-        yield 'loosely' => ['loosely'];
+        yield 'loose' => ['loose'];
+    }
+
+    /**
+     * @test
+     * @dataProvider provideMigrationValues
+     */
+    public function deprecatedValueIsProperlyMigrated(string $existingValue, string $expectedResult): void
+    {
+        self::assertEquals($expectedResult, $this->configOption->tryToMigrateValue($existingValue));
+    }
+
+    public function provideMigrationValues(): iterable
+    {
+        yield 'strict' => ['strict', 'strict'];
+        yield 'loose' => ['loose', 'loose'];
+        yield 'loosely' => ['loosely', 'loose'];
     }
 }